first stage of staging changes

+++ snapshot2/snapshot_prod-rsync.sh
@@ -1,10 +1,19 @@
-/bin/bash
+#!/bin/bash
+
+###snapshot_prod-rsync.sh <PRODDB> <SSHSTAGING> <SCRIPTDIR>
 
 set -uex
+[ ! -f ${CWD}/conf ] && exit 1
+source /etc/dumpdb/conf
+
+DUMPDIR="${PRODDUMPDIR}/${PRODDB}"
+DESTINATIONDUMP="${STAGINGDEST}/${PRODDB}"
+
+
 PRODDB="${1}"
-DUMPDIR="${2}/${PRODDB}"
-DESTINATIONDUMP="${3}"
-SSHUSER=$"{4}"
+SSHSTAGING="${2}"
+CWD="${3}"
+
 rsync -zavhP --delete --exclude-from ${CWD}/rsync-table-exclusions.txt \
-  ${DUMPDIR} ${SSHUSER}@{$SSHDESTINATIONDUMP}:${DESTINATIONDUMP}/
+  ${DUMPDIR} ${SSHSTAGING}:${DESTINATIONDUMP}/
